Danny Test Changes
Changes that I have so far trying to extract data out of the
statisticsTracker component. Looking more at the scripting code I can
see that making a new c++ class for the component is not the right
approach after running into problems. Look at the GuiInterface.js file
at the GetExtendedSimulationState function for a better approach. My new
thought is that we can use this function as an example to add new
functions on the script side to get an array of data that we want to a
component like playerManager that is already set up on the C side. This
push is more for me to be able to get what I have on my desktop because
my laptop keeps acting up and preventing me from working not so much
because I need help. But if anyone wants to try and accomplish what I
described feel free.