Prototype Code for Player State Tables
This is a preliminary commit with the foundations of the code for
extracting all relevant data from the statistics tracker component.

This builds a state table for each player in the game and has a method
for extracting state vectors containing 23 features. This one includes
all of the data we originally talked about plus some other stuff I
thought might be relevant. Obviously we can change the features if we
need to.

This probably has some quirks that will need to be worked out. I have
not tested this yet so there may be issues with how the 3d vector is
coded. As far as data extraction goes I'm nearly certain this is
correct. I tested out the method with only getting the players unit's
trained and was getting good data written to the file. This commit just
expands on the previous one to get more data and store it in a usable
vector for the other tasks.

Once again this is a commit that is more for me so that I can access my
work when I get back to Lawrence tonight.