Added Linux sockets
I tried to add multiplatform sockets, after telling Katherine I'm pretty
sure it would work. Bad news, it didn't. But I was able to get the linux
sockets working. I tried to get winsock sockets working too, since it
was my fault, but I kept getting redefinition errors (the same problem I
was having when using the multiplatform boost asio library).
Technically, I haven't tested this, but everything on windows works
still. I needed to upload this to get it on a box. Also, David, we are
going to need to talk about some server changes that need made, because
if we want live data streaming, we are going to have to read on multiple
ports server-side.